src/requirements.txt
duckdb>=1.2.0
google-genai
langchain-text-splitters
langchain-groq
python-dotenv
typer[all]
numpy
rich
fastapi
uvicorn
python-multipart
easyocr
vosk
imageio-ffmpeg
pdfplumber
yt-dlp